backend/requirements.txt
openai==1.3.0
mistralai==1.0.0
httpx>=0.27.0,<0.28.0
python-dotenv==1.0.0
